move isolayer logic to separate mixin-like class
